move all code related to generating report to reporter.py
