+++ gentoo_update/__init__.py
@@ -1,4 +1,5 @@
 from .gentoo_update import add_prefixes, create_cli, main
 from .shell_runner import ShellRunner
 from .parser import Parser
+from .reporter import Reporter
 from ._version import __version__
